Refactor getting and writing settings
Reading from the config file every time a setting was accessed meant a
lot of configparser.ConfigParser() objects being created and a lot of
opening of settings.ini. Retrieving all of the settings into a
dictionary and accessing and modifying that as and when needed should be
less taxing and reduces the amount of mocking needed for the tests.